Split drop target transfer worker
